move in-memory peerstore into mem package.
