+++ ds/peerstore.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/ipfs/go-datastore"
 	pstore "github.com/libp2p/go-libp2p-peerstore"
+	"github.com/libp2p/go-libp2p-peerstore/mem"
 )
 
 // NewPeerstore creates a peerstore backed by the provided persistent datastore.
@@ -15,6 +16,6 @@ func NewPeerstore(ctx context.Context, ds datastore.Batching) (pstore.Peerstore,
 		return nil, err
 	}
 
-	ps := pstore.NewPeerstoreWith(pstore.NewKeybook(), addrBook, pstore.NewPeerMetadata())
+	ps := pstore.NewPeerstoreWith(mem.NewKeybook(), addrBook, mem.NewPeerMetadata())
 	return ps, nil
 }
