+++ mem/inmem_test.go
@@ -1,25 +1,26 @@
-package peerstore
+package mem
 
 import (
 	"testing"
 
+	pstore "github.com/libp2p/go-libp2p-peerstore"
 	"github.com/libp2p/go-libp2p-peerstore/test"
 )
 
 func TestInMemoryPeerstore(t *testing.T) {
-	test.TestPeerstore(t, func() (Peerstore, func()) {
+	test.TestPeerstore(t, func() (pstore.Peerstore, func()) {
 		return NewPeerstore(), nil
 	})
 }
 
 func TestInMemoryAddrMgr(t *testing.T) {
-	test.TestAddrMgr(t, func() (AddrBook, func()) {
+	test.TestAddrMgr(t, func() (pstore.AddrBook, func()) {
 		return &AddrManager{}, nil
 	})
 }
 
 func BenchmarkInMemoryPeerstore(b *testing.B) {
-	test.BenchmarkPeerstore(b, func() (Peerstore, func()) {
+	test.BenchmarkPeerstore(b, func() (pstore.Peerstore, func()) {
 		return NewPeerstore(), nil
 	})
 }
