+++ mem/peerstore.go
@@ -1,0 +1,11 @@
+package mem
+
+import pstore "github.com/libp2p/go-libp2p-peerstore"
+
+// NewPeerstore creates an in-memory threadsafe collection of peers.
+func NewPeerstore() pstore.Peerstore {
+	return pstore.NewPeerstoreWith(
+		NewKeybook(),
+		&AddrManager{},
+		NewPeerMetadata())
+}
